refactor: move querycursor under the cursor dir
