+++ lib/cursor/QueryCursor.js
@@ -2,9 +2,9 @@
  * Module dependencies.
  */
 
-var PromiseProvider = require('./promise_provider');
+var PromiseProvider = require('../promise_provider');
 var Readable = require('stream').Readable;
-var helpers = require('./queryhelpers');
+var helpers = require('../queryhelpers');
 var util = require('util');
 
 /**
