+++ lib/query.js
@@ -3,7 +3,7 @@
  */
 
 var PromiseProvider = require('./promise_provider');
-var QueryCursor = require('./querycursor');
+var QueryCursor = require('./cursor/QueryCursor');
 var QueryStream = require('./querystream');
 var cast = require('./cast');
 var castUpdate = require('./services/query/castUpdate');
